+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/KPromiseWrapper.kt
@@ -3,22 +3,18 @@ package expo.modules.kotlin
 import android.os.Bundle
 import com.facebook.react.bridge.Arguments
 
-class KotlinPromiseWrapper(
+class KPromiseWrapper(
   private val bridgePromise: com.facebook.react.bridge.Promise
 ) : Promise {
 
   override fun resolve(value: Any?) {
-    when (value) {
-      is Bundle -> {
-        bridgePromise.resolve(Arguments.fromBundle(value as Bundle?))
+    bridgePromise.resolve(
+      when (value) {
+        is Bundle -> Arguments.fromBundle(value as Bundle?)
+        is List<*> -> Arguments.fromList(value as List<*>?)
+        else -> value
       }
-      is List<*> -> {
-        bridgePromise.resolve(Arguments.fromList(value as List<*>?))
-      }
-      else -> {
-        bridgePromise.resolve(value)
-      }
-    }
+    )
   }
 
   override fun reject(code: String, message: String?, cause: Throwable?) {
