--- packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/exception/InvalidArgsNumberException.kt
@@ -1,4 +1,0 @@
-package expo.modules.kotlin.exception
-
-class InvalidArgsNumberException(received: Int, expected: Int)
-  : CodedException(message = "Received $received arguments, but $expected was expected.")
