--- packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/exception/MethodNotFoundException.kt
@@ -1,4 +1,0 @@
-package expo.modules.kotlin.exception
-
-internal class MethodNotFoundException(methodName: String, moduleName: String)
-  : CodedException(message = "Cannot fund method $methodName in module $moduleName")
