--- packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/exception/NullArgumentException.kt
@@ -1,6 +1,0 @@
-package expo.modules.kotlin.exception
-
-import kotlin.reflect.KType
-
-class NullArgumentException(desiredType: KType) :
-  CodedException(message = "Cannot assigned null to not nullable type $desiredType")
