--- packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/exception/UnexpectedException.kt
@@ -1,4 +1,0 @@
-package expo.modules.kotlin.exception
-
-class UnexpectedException(val throwable: Throwable)
-  : CodedException(throwable)
