+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/methods/Method.kt
@@ -8,7 +8,7 @@ class Method(
   argsType: Array<KType>,
   private val body: (args: Array<out Any?>) -> Any?
 ) : AnyMethod(name, argsType) {
-  override fun callUserImplementation(args: Array<out Any?>, promise: Promise) {
+  override fun callImplementation(args: Array<out Any?>, promise: Promise) {
     promise.resolve(body(args))
   }
 }
